Extracting styles and WorkspaceEditComponent seperate file
